Drop the index when writing the CSV file
The index is a Pandas RangeIndex [1], which implements a monotonic
integer range and was added when the CSV file was read; it wasn't added
when the CSV file was written by cohort-extractor. By dropping it when
writing the CSV file, we ensure the CSV file has the same shape as a CSV
file that was written by cohort-extractor.

[1]: https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.RangeIndex.html